Replace the Nomis::Api namespace with Nomis::Elite2
