+++ app/services/movement_service.rb
@@ -1,7 +1,7 @@
 class MovementService
   # rubocop:disable Metrics/MethodLength
   def self.movements_on(date, direction_filters: [], type_filters: [])
-    movements = Nomis::Api::MovementApi.movements_on_date(date)
+    movements = Nomis::Elite2::MovementApi.movements_on_date(date)
 
     if direction_filters.any?
       movements = movements.select { |m|
