+++ app/services/nomis/elite2/elite2_api.rb
@@ -1,11 +1,11 @@
 module Nomis
-  module Api
+  module Elite2
     ApiPaginatedResponse = Struct.new(:meta, :data)
 
     module Elite2Api
       def e2_client
         host = Rails.configuration.nomis_oauth_host
-        Nomis::Client.new(host)
+        Nomis::Elite2::Client.new(host)
       end
 
       def api_deserialiser
