+++ app/services/prison_offender_manager_service.rb
@@ -7,7 +7,7 @@ class PrisonOffenderManagerService
   end
 
   def self.get_poms(prison)
-    poms = Nomis::Api::PrisonOffenderManagerApi.list(prison)
+    poms = Nomis::Elite2::PrisonOffenderManagerApi.list(prison)
 
     poms = poms.map { |pom|
       detail = get_pom_detail(pom.staff_id)
@@ -56,7 +56,7 @@ class PrisonOffenderManagerService
   end
 
   def self.get_signed_in_pom_details(current_user)
-    user = Nomis::Api::UserApi.user_details(current_user)
+    user = Nomis::Elite2::UserApi.user_details(current_user)
     poms_list = PrisonOffenderManagerService.get_poms(user.active_case_load_id)
     @pom = poms_list.select { |p| p.staff_id.to_i == user.staff_id.to_i }.first
   end
